requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langgraph
chromadb
sentence-transformers
duckduckgo-search
streamlit
python-dotenv
beautifulsoup4
requests